Split Events into EventSource and Subscription, add docs
